21-10-24

+++ web/Server/api.js
@@ -1,6 +1,8 @@
 
 var express = require('express')
-const {Mint, ClientAccountID, Transfer,Request, TotalBids, GetBidsByIndex,Register,Login, GetAccountBalance} = require("./fabric");
+const {Mint, ClientAccountID, Transfer,Request, TotalBids, GetBidsByIndex,Register,Login, GetAccountBalance, TotalNFTs,
+    GetNFTByIndex, IsOnSale, AddBid, Offer, IsNFTExist
+} = require("./fabric");
 var router = express.Router()
 
 const multer = require('multer')
@@ -32,7 +34,6 @@ router.post('/register',async(req,res)=>{
         res.status(200).send(result)
     }
     catch(err){
-        console.log(err)
         res.status(404).send(err)
     }
 })
@@ -45,7 +46,6 @@ router.post('/login',async(req,res)=>{
         res.status(200).send('Login Successfully')
     }
     catch(err){
-        console.log(`******** FAILED to Login: ${err}`)
         res.status(404).send(err)
     }
 })
@@ -58,22 +58,36 @@ router.post('/mint', upload.single('file'), async (req, res, next) => {
     console.log(req.body.org)
     console.log(req.body.tokenID)
     console.log(req.file)
+    const clientID=req.body.clientID
+    const org=req.body.org
+    const tokenID=req.body.tokenID
     // 获取文件
-    let file = req.file
-    let newname
-    if (file) {
-        // 获取文件名
-        let fileNameArr = req.body.tokenID+'.png'
-        // 获取文件后缀名
-        var suffix = fileNameArr[fileNameArr.length - 1]
-        // 文件重命名
-        newname=`./uploads/${fileNameArr}`
-        fs.renameSync('./uploads/' + file.filename, newname)
-        file['filename'] = `${fileNameArr}`
-    }
+    try{
+            const exist = await IsNFTExist(clientID,org,tokenID)
+            if( exist.toString()==='true'){
+                res.status(200).send('tokenID: '+ tokenID + ' already exists')
+            }else if (exist.toString()==='false'){
+                let file = req.file
+                let newname
+                if (file) {
+                    // 获取文件名
+                    let fileNameArr = tokenID+'.png'
+                    // 获取文件后缀名
+                    var suffix = fileNameArr[fileNameArr.length - 1]
+                    // 文件重命名
+                    newname=`./uploads/${fileNameArr}`
+                    fs.renameSync('./uploads/' + file.filename, newname)
+                    file['filename'] = `${fileNameArr}`
+                }
 
-    let result=await Mint(req.body.clientID, req.body.org, req.body.tokenID, newname)
-    res.status(200).send('Successfully Mint NFT with id: ' + req.body.tokenID + ' for: ' + req.body.clientID + ', CID: '+result.toString())
+                let result=await Mint(clientID, org, tokenID, newname)
+                res.status(200).send('Successfully Mint NFT with id: ' + tokenID + ' for: ' + clientID + ', CID: '+result.toString())
+            }else{
+                res.status(404)
+            }
+    }catch(err){
+        res.status(404).send(err)
+    }
 })
 
 
@@ -91,17 +105,43 @@ router.post('/account',async (req,res)=>{
     let org=param.get('org')
     console.log(clientid)
     console.log(org)
-    let result=await GetAccountBalance(clientid,org)
-    console.log(result)
-    res.status(200).send('user: ' + clientid + '\nbalance: '+ result.Balance +'\nencoded accountID: ' + result.Account)
+    try{
+        let result=await GetAccountBalance(clientid,org)
+        console.log(result)
+        res.status(200).send('user: ' + clientid + '\nbalance: '+ result.Balance +'\nencoded accountID: ' + result.Account)
+    }catch(err){
+        res.status(404).send(err)
+    }
+
+})
+router.post('/totalnfts',async(req,res)=>{
+    console.log(req.url)
+    let param = US(req.url)
+    let clientid=param.get('clientID')
+    let org=param.get('org')
+    try{
+        let result=await TotalNFTs(clientid,org)
+        res.status(200).send(result.toString())
+    }
+    catch (err) {
+        console.log(err)
+        res.status(404)
+    }
 })
 router.post('/totalbids',async(req,res)=>{
     console.log(req.url)
     let param = US(req.url)
     let clientid=param.get('clientID')
     let org=param.get('org')
-    let result=await TotalBids(clientid,org)
-    res.status(200).send(result.toString())
+    try{
+        let result=await TotalBids(clientid,org)
+        console.log(result)
+        res.status(200).send(result.toString())
+    }
+    catch (err) {
+        console.log(err)
+        res.status(404)
+    }
 })
 
 router.post('/getBidsByIndex',async(req,res)=>{
@@ -110,10 +150,73 @@ router.post('/getBidsByIndex',async(req,res)=>{
     let org=param.get('org')
     let index=param.get('index')    //string
     console.log(req.url)
-    let result=await GetBidsByIndex(clientid,org,index)
-    let parsedresult=JSON.parse(result.toString())
-    res.status(200).send(parsedresult)
+    try{
+        let result=await GetBidsByIndex(clientid,org,index)
+        let parsedresult=JSON.parse(result.toString())
+        //console.log(parsedresult)
+        res.status(200).send(parsedresult)
+    }
+    catch(err){
+        res.status(404)
+    }
+
 })
+
+router.post('/getNFTByIndex',async(req,res)=>{
+    console.log(req.url)
+    let param = US(req.url)
+    let clientid=param.get('clientID')
+    let org=param.get('org')
+    let index=param.get('index')    //string
+    try{
+        let result=await GetNFTByIndex(clientid,org,index)
+        let parsedresult=JSON.parse(result.toString())
+
+        let onSale = await IsOnSale(clientid,org,parsedresult.ID)
+        let jsonResult= {'tokenID':parsedresult.ID, 'CID':parsedresult.CID, 'Status':onSale.toString()}
+        res.status(200).send(jsonResult)
+    }
+    catch(err){
+        res.status(404)
+    }
+
+})
+router.post('/addbid',async(req,res)=>{
+    console.log(req.url)
+    let param = US(req.url)
+    let clientid=param.get('clientID')
+    let org=param.get('org')
+    let tid=param.get('tokenID')
+    let lp=param.get('lowerPrice')
+    let hp=param.get('higherPrice')
+    let lt=param.get('lifeTime')
+    try{
+        let result=await AddBid(clientid,org,tid,lp,hp,lt)
+        let parsedresult=JSON.parse(result.toString())
+        res.status(200).send(parsedresult)
+    }
+    catch(err){
+        res.status(404).send(err)
+    }
+})
+
+router.post('/offer',async(req,res)=>{
+    console.log(req.url)
+    let param = US(req.url)
+    let clientid=param.get('clientID')
+    let org=param.get('org')
+    let tid=param.get('tokenID')
+    let price=param.get('price')
+    try{
+        await Offer(clientid,org,tid,price)
+        res.status(200).send('successfully offer')
+    }
+    catch(err){
+        res.status(404).send(err)
+    }
+})
+
+
 router.post('/transfer',async (req,res)=>{
     let param = US(req.url)
     let clientid=param.get('clientID')
@@ -124,9 +227,13 @@ router.post('/transfer',async (req,res)=>{
     console.log(org)
     console.log(tokenid)
     console.log(targetid)
-    let result=await Transfer(clientid,org,tokenid,targetid)
-    console.log(result)
-    res.status(200).send('transfer token '+tokenid+' '+result)
+    try{
+        let result=await Transfer(clientid,org,tokenid,targetid)
+        console.log(result)
+        res.status(200).send('transfer token '+tokenid+' '+result)
+    }catch(err){
+        res.status(404).send(err)
+    }
 })
 
 router.post('/user', function (req, res) {
